pyproject.toml
[tool.poetry]
name = "fastapi-tutorial"
version = "0.1.0"
description = "fastapi-tutorial"
authors = ["davesanna <sanna.davide32@gmail.com>"]
license = "MIT"
readme = "README.md" 

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.0.0"
pandas = "^1.4.0"
fastapi = {extras = ["all"], version = "^0.73.0"}
psycopg2-binary = "^2.9.3"
flake8 = "^4.0.1"
black = "^22.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg = "^3.0.9"
SQLAlchemy = "^1.4.31"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
mypy = "^0.931"
black = "^22.1.0"
isort = "^5.10.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fastapi-tutorial = "fastapi.main:app"